+++ terraform/pipeline/step-functions/dynamic/Ind-CQC-Filled-Post-Estimates.json
@@ -149,8 +149,39 @@
                   }
                 },
                 {
-                  "StartAt": "Polars Impute IND CQC data",
+                  "StartAt": "Polars prepare impute IND CQC data",
                   "States": {
+                    "Polars prepare impute IND CQC data": {
+                      "Type": "Task",
+                      "Resource": "arn:aws:states:::ecs:runTask.sync",
+                      "Parameters": {
+                        "Cluster": "${polars_cluster_arn}",
+                        "TaskDefinition": "${independent_cqc_task_arn}",
+                        "LaunchType": "FARGATE",
+                        "NetworkConfiguration": {
+                          "AwsvpcConfiguration": {
+                            "Subnets": ${public_subnet_ids},
+                            "SecurityGroups": ["${independent_cqc_security_group_id}"],
+                            "AssignPublicIp": "ENABLED"
+                          }
+                        },
+                        "Overrides": {
+                          "ContainerOverrides": [
+                            {
+                              "Name": "_03_independent_cqc-container",
+                              "Command": [
+                                "impute_prepare.py",
+                                "--cleaned_ind_cqc_source",
+                                "${dataset_bucket_uri}/domain=ind_cqc_filled_posts/dataset=ind_cqc_02_cleaned_data/",
+                                "--destination",
+                                "${dataset_bucket_uri}/domain=ind_cqc_filled_posts/dataset=ind_cqc_03_impute_prepared/"
+                              ]
+                            }
+                          ]
+                        }
+                      },
+                      "Next": "Polars Impute IND CQC data"
+                    },
                     "Polars Impute IND CQC data": {
                       "Type": "Task",
                       "Resource": "arn:aws:states:::ecs:runTask.sync",
@@ -172,7 +203,7 @@
                               "Command": [
                                 "impute_ind_cqc_ascwds_and_pir.py",
                                 "--cleaned_ind_cqc_source",
-                                "${dataset_bucket_uri}/domain=ind_cqc_filled_posts/dataset=ind_cqc_02_cleaned_data/",
+                                "${dataset_bucket_uri}/domain=ind_cqc_filled_posts/dataset=ind_cqc_03_impute_prepared/",
                                 "--destination",
                                 "${dataset_bucket_uri}/domain=ind_cqc_filled_posts/dataset=ind_cqc_03_imputed_ascwds_and_pir_polars/"
                               ]
